+++ CHANGELOG.md
@@ -1,3 +1,40 @@
+# [1.396.0](https://github.com/adobe/spacecat-api-service/compare/v1.395.0...v1.396.0) (2026-04-01)
+
+
+### Features
+
+* add created by + created at fields in llmo config response; fix missing updated at in category ([#2089](https://github.com/adobe/spacecat-api-service/issues/2089)) ([616fde8](https://github.com/adobe/spacecat-api-service/commit/616fde89ac9dc875e326083f598df8d19614fcd7))
+
+# [1.395.0](https://github.com/adobe/spacecat-api-service/compare/v1.394.1...v1.395.0) (2026-04-01)
+
+
+### Features
+
+* adds token controllers ([#1941](https://github.com/adobe/spacecat-api-service/issues/1941)) ([e5383fa](https://github.com/adobe/spacecat-api-service/commit/e5383fa19fb2b054395e939429c5b140ab591dcb))
+* use new brand-presence filter dimensions rpc ([#2077](https://github.com/adobe/spacecat-api-service/issues/2077)) ([04f5339](https://github.com/adobe/spacecat-api-service/commit/04f53390dbde4071df6a3d5cbf24e93cef838cd1))
+
+## [1.394.1](https://github.com/adobe/spacecat-api-service/compare/v1.394.0...v1.394.1) (2026-04-01)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#2088](https://github.com/adobe/spacecat-api-service/issues/2088)) ([7c0aa27](https://github.com/adobe/spacecat-api-service/commit/7c0aa27159b85eef0dff77a6b59af68f6e456c63))
+
+# [1.394.0](https://github.com/adobe/spacecat-api-service/compare/v1.393.0...v1.394.0) (2026-04-01)
+
+
+### Features
+
+* metadata api ([#2068](https://github.com/adobe/spacecat-api-service/issues/2068)) ([e8de8a1](https://github.com/adobe/spacecat-api-service/commit/e8de8a17d937315359d3e839d70b27f5703dbc0c))
+
+# [1.393.0](https://github.com/adobe/spacecat-api-service/compare/v1.392.0...v1.393.0) (2026-04-01)
+
+
+### Features
+
+* log dryrun set to false | LLMO-3918 ([#2090](https://github.com/adobe/spacecat-api-service/issues/2090)) ([5569d18](https://github.com/adobe/spacecat-api-service/commit/5569d18a19e674a27342181c5102b36df8bffc23))
+* no forced dryRun for LLMO config sync | LLMO-3918 ([#2085](https://github.com/adobe/spacecat-api-service/issues/2085)) ([b3caf3b](https://github.com/adobe/spacecat-api-service/commit/b3caf3b100f1d41ad482d7b4963f05f8bd832703))
+
 # [1.392.0](https://github.com/adobe/spacecat-api-service/compare/v1.391.0...v1.392.0) (2026-03-31)